fix: avoid issues with strings in exec commands
Commands passed as override_exec when the user clicks an app menu
entry like the following

   qvm-run -q work 'touch note.txt'

where being split into

   ["qvm-run", "-q", "work", "'touch", "note.txt'"]

instead of

   ["qvm-run", "-q", "work", "'touch note.txt'"]

To avoid issues like this override_exec commands are called
unsecurely (since the tutorial file is assumed trusted) as a shell
command.

'true' is then passed on as a return command from the decorated
'get_command_for_vm()' just so it has something to run.